Adding examples: testing a non-collaborative policy
Policies are supposed to tell the outside world the outcome of their
promises by means of reports (useful for humans and for log analysis
tools) and globally scoped classes (useful to the agent running the
policies).

If a policy fails to set classes to report the agent about its outcomes,
it may still be possible to tell what it's doing by examining the
outcome of the actions the policy takes.

10-file_make.t is an example of such a test. The bundles being tested
are part of the standard library shipped with CFEngine Community 3.7.3.
The files are included here for convenience but are *not* part of
the cfengine-tap distribution. Please refer to CFEngine's official
masterfiles for them.