Adding examples: testing a collaborative policy
Policies are supposed to tell the outside world the outcome of their
promises by means of reports (useful for humans and for log analysis
tools) and globally scoped classes (useful to the agent running the
policies).

The bundle agent check_process_running included in the file
process_example.cf is an example of such a policy. The policy takes
a string or regular expression as a parameter and checks the
process table for a match.  The policy sets a global class, whose name
is computed from the process name, to inform the agent about the
outcome. E.g., if you check for /sbin/init, the bundle will set:

* process__sbin_init_running_ok if the process is found running;
* process__sbin_init_running_not_kept if the process is not found.

This is similar to the body classes classes_generic included in the
standard library.

The test policy in 20-process.t will check, first and foremost,
if you have /sbin/init on your system. If not, the first test will
fail, which will make the whole script appear as failed.

It will then check if /sbin/init is found in the process table. It
expects it to be there and will consider the test passed if it finds
it.

Finally, it will look for a fake process (which I named "starvestone")
and expect not to find it. If the process is not running (and I really
hope it's not!) the test will succeed.